--- client/src/components/homeOtros/Perfil.jsx
@@ -1,29 +1,0 @@
-import React from "react";
-import { teal } from "@mui/material/colors";
-import { ModeEditOutlined } from "@mui/icons-material";
-import { Grid, IconButton } from "@mui/material";
-import { useStyles } from "../../styles/otroHome/perfil";
-import circleUser from "../../img/user.png";
-import BodyData from "./perfil/BodyData";
-
-function Perfil() {
-  const classes = useStyles();
-
-  return (
-    <Grid container item className={classes.root}>
-      <IconButton
-        style={{
-          position: "absolute",
-          top: ".5em",
-          right: ".5em",
-        }}
-      >
-        <ModeEditOutlined fontSize="medium" style={{ color: teal[800] }} />
-      </IconButton>
-      <img src={circleUser} alt="user profile" className={classes.circleUser} />
-      <BodyData classes={classes} />
-    </Grid>
-  );
-}
-
-export default Perfil;
