+++ client/src/styles/home/index.js
@@ -1,7 +1,12 @@
 import { makeStyles } from "@material-ui/core/styles";
+import { teal } from "@mui/material/colors";
 
 export const useStyles = makeStyles({
   root: {
+    height: "100vh",
+    background: teal[50],
+  },
+  container: {
     height: "90vh",
     margin: "0 1em",
     alignItems: "flex-end",
